Maximize potential of S token payouts through dynamic auctions
This is a major commit where:
1. We have set a max limit of 1000 shield markets per loan market (due to fixed array size restriction in Vyper).
2. With every purchase, the settlement price (currency raised / underlying sold) changes and thus, the maximum S token payout amount for the corresponding loan market is re-calculated.
3. This recalculation reduces the amount of currency required to be raised during the auction.

+++ contracts/daos/ProtocolDao.v.py
@@ -511,7 +511,7 @@ def set_expiry_support(_timestamp: timestamp, _label: string[3], _is_active: boo
 
 
 @public
-def set_registry(_dao_type: int128, _registry_type: uint256, _address: address) -> bool:
+def set_registry(_dao_type: int128, _registry_type: int128, _address: address) -> bool:
     """
         @dev Function to set / change the Registry address across the system.
              Only the Governor can call this function.
